+++ src/main/kotlin/lottery/Lottery.kt
@@ -1,0 +1,9 @@
+package lottery
+
+class Lottery(val lotteryNumbers: LotteryNumbers) {
+    companion object {
+        fun create(): Lottery {
+            return Lottery(LotteryNumbers.create())
+        }
+    }
+}
