+++ src/main/kotlin/lottery/LotteryMachine.kt
@@ -1,0 +1,9 @@
+package lottery
+
+class LotteryMachine {
+    companion object {
+        fun buy(count: Int): List<Lottery> {
+            return List(count) { Lottery.create() }
+        }
+    }
+}
